cleanup test directory
