+++ tools/tests/py2fgen/test_codegen_references.py
@@ -15,6 +15,9 @@ from click.testing import CliRunner
 from icon4py.tools.py2fgen import _cli
 
 
+# TODO(havogt): move to `wrappers` test
+
+
 @pytest.fixture
 def cli_runner():
     return CliRunner()
