+++ tools/tests/py2fgen/test_generator.py
@@ -6,38 +6,28 @@
 # Please, refer to the LICENSE file in the root directory.
 # SPDX-License-Identifier: BSD-3-Clause
 
+import pathlib
 import subprocess
 import tempfile
-from pathlib import Path
 
-import numpy as np
+import cffi
 import pytest
-from cffi import FFI
 
-from icon4py.tools.py2fgen._conversion import _unpack_numpy
-from icon4py.tools.py2fgen._generator import generate_and_compile_cffi_plugin
+from icon4py.tools.py2fgen._codegen import CffiPlugin
+from icon4py.tools.py2fgen._generator import generate_and_compile_cffi_plugin, get_cffi_description
+
+
+def test_parse_functions_on_wrapper():
+    # TODO make indepedent of `wrappers`
+    module_path = "icon4py.tools.py2fgen.wrappers.diffusion_wrapper"
+    functions = ["diffusion_init", "diffusion_run"]
+    plugin = get_cffi_description(module_path, functions, "diffusion_plugin")
+    assert isinstance(plugin, CffiPlugin)
 
 
 @pytest.fixture
 def ffi():
-    return FFI()
-
-
-@pytest.mark.parametrize(
-    "data, expected_result",
-    [
-        ([1.0, 2.0, 3.0, 4.0], np.array([[1.0, 3.0], [2.0, 4.0]])),
-        ([1, 2, 3, 4], np.array([[1, 3], [2, 4]])),
-    ],
-)
-def test_unpack_column_major(data, expected_result, ffi):
-    ptr = ffi.new("double[]", data) if isinstance(data[0], float) else ffi.new("int[]", data)
-
-    rows, cols = expected_result.shape
-
-    result = _unpack_numpy(ffi, ptr, rows, cols)
-
-    assert np.array_equal(result, expected_result)
+    return cffi.iFFI()
 
 
 def test_compile_and_run_cffi_plugin_from_C():
@@ -62,7 +52,7 @@ def test_compile_and_run_cffi_plugin_from_C():
     """
 
     with tempfile.TemporaryDirectory() as tmpdirname:
-        build_path = Path(tmpdirname)
+        build_path = pathlib.Path(tmpdirname)
 
         try:
             # Generate and compile the CFFI plugin, which creates lib{plugin_name}.so
